# swagger/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "CryptoPill",
		"contact": {},
		"version": "1.0",
		"description": "API Docs"
	},
	"servers": [
		{
			"url": "http://180.149.241.208:3048",
			"variables": {}
		},
		{
			"url": "http://localhost:3044",
			"variables": {}
		}
	],
	"paths": {
...
		"/user/forgotPassword/{ph_number}": {
			"get": {
				"tags": ["User"],
				"summary": "ForgotPassword",
				"operationId": "ForgotPassword",
				"parameters": [
					{
						"name": "ph_number",
						"in": "path",
						"description": "",
						"required": true,
						"style": "simple",
						"schema": {
							"type": "string",
							"example": "+919594521511"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/user/login": {
			"post": {
				"tags": ["User"],
				"summary": "Step 4: Login",
				"operationId": "Login",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							},
							"example": {
								"contact_number": "+918178426170",
								"password": "12345678"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponse"
								},
								"example": {
									"success": true,
									"data": {
										"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZmlyc3RfbmFtZSI6InNhdGlzaCIsIm1pZGRsZV9uYW1lIjoiIiwibGFzdF9uYW1lIjoiam9zaGkiLCJlbWFpbCI6ImR1bW15ZG9jQGdtYWlsLmNvbSIsImdlbmRlciI6Im1hbGUiLCJiaXJ0aF9kYXRlIjoiMTk4NS0wOC0wMSIsInByb2ZpbGVfaW1hZ2UiOiJodHRwczovL2NyeXB0b3BpbGxhcHAuczMuYW1hem9uYXdzLmNvbS9kb2N0b3IvMS9kZW1vLmpwZyIsIm5ld19wcm9maWxlX2ltYWdlIjpudWxsLCJjb250YWN0X251bWJlciI6Iis5MTgxNzg0MjYxNzAiLCJhY2NvdW50X2FjdGl2YXRpb24iOjEsInByb2ZpbGVfaW1hZ2VfdmVyaWZ5IjowLCJsYWJfb3JfcGhhcm1hX2VtcGxveWVtZW50X251bWJlciI6bnVsbCwiZW1haWxfdmVyaWZ5IjowLCJwaG9uZV92ZXJpZnkiOjEsImRlZmF1bHRfcm9sZSI6MiwiaXNBZG1pbiI6MCwiaXNTZXR1cENvbXBsZXRlIjoxLCJkZXZpY2VUb2tlbiI6bnVsbCwiZGV2aWNlVHlwZSI6bnVsbCwicm9sZV9pZCI6Miwicm9sZSI6ImRvY3RvciIsImlzV29ya3BsYWNlQWRtaW4iOjEsImlhdCI6MTYwNTY4MDA5MSwiZXhwIjoxNjA2OTc2MDkxfQ.x35_SNQCcEIRHt2_fmj-8-hxZ3YGeyiLRTt9B87sGrk",
										"default_role": 2,
										"id": 1,
										"isSetupComplete": 1
									},
									"message": "Success"
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/user/updatePassword": {
			"post": {
				"tags": ["User"],
				"summary": "Step 3: UpdatePassword",
				"operationId": "UpdatePassword",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePasswordRequest"
							},
							"example": {
								"id": 1,
								"password": "12345678"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/user/verifyOtp": {
			"post": {
				"tags": ["User"],
				"summary": "Step 2: VerifyOTP",
				"operationId": "VerifyOTP",
				"parameters": [
					{
						"name": "Origin",
						"in": "header",
						"description": "",
						"required": true,
						"style": "simple",
						"schema": {
							"type": "string",
							"example": "localhost:3000"
						}
					}
				],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyOTPRequest"
							},
							"example": {
								"ph_number": "+918178426171",
								"otp": "570605"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/user/requestOTP": {
			"post": {
				"tags": ["User"],
				"summary": "Step 1: Request OTP",
				"operationId": "RequestOTP",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestOTPRequest"
							},
							"example": {
								"first_name": "patient",
								"role_id": 2,
								"middle_name": "",
								"last_name": "user",
								"contact_number": "+918178426171"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/user/changePassword": {
			"post": {
				"tags": ["User"],
				"summary": "Change Password",
				"operationId": "ChangePassword",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordRequest"
							},
							"example": {
								"current_password": "krishank1234",
								"new_password": "krishank@123",
								"confirm_new_password": "krishank@123"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
		"/user/resendOTP": {
			"post": {
			  "tags": [
				"User"
			  ],
			  "summary": "ResendOTP",
			  "operationId": "ResendOTP",
			  "parameters": [],
			  "requestBody": {
				"description": "",
				"content": {
				  "application/json": {
					"schema": {
					  "$ref": "#/components/schemas/ResendOTPRequest"
					},
					"example": {
					  "contact_number": "+918128153228"
					}
				  }
				},
				"required": true
			  },
			  "responses": {
				"200": {
				  "description": "",
				  "headers": {}
				}
			  },
			  "deprecated": false,
			  "security": []
			}
		},
		"/user/removeuser/{contact_number}": {
			"get": {
			  "tags": [
				"User"
			  ],
			  "summary": "removeuser",
			  "operationId": "removeuser",
			  "parameters": [
				{
				  "name": "contact_number",
				  "in": "path",
				  "description": "",
				  "required": true,
				  "style": "simple",
				  "schema": {
					"type": "string",
					"example": "+918087618736"
				  }
				}
			  ],
			  "responses": {
				"200": {
				  "description": "",
				  "headers": {}
				}
			  },
			  "deprecated": false,
			  "security": []
			}
		},
...
		"/file/uploadProfile": {
			"post": {
				"tags": ["File"],
				"summary": "uploadProfile",
				"operationId": "uploadProfile",
				"parameters": [],
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"encoding": {},
							"schema": {
								"required": ["files", "role_id", "user_id"],
								"type": "object",
								"properties": {
									"files": {
										"type": "string",
										"format": "binary"
									},
									"role_id": {
										"type": "integer",
										"format": "int32",
										"example": 4
									},
									"user_id": {
										"type": "integer",
										"format": "int32",
										"example": 2
									}
								}
							}
						}
					},
					"required": false
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"LoginRequest": {
				"title": "Login Request",
				"required": ["contact_number", "password"],
				"type": "object",
				"properties": {
					"contact_number": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"example": {
					"contact_number": "+918178426170",
					"password": "12345678"
				}
			},
			"LoginResponse": {
				"title": "Login Response",
				"required": ["success", "data", "message"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/Data2"
					},
					"message": {
						"type": "string"
					}
				},
				"example": {
					"success": true,
					"data": {
						"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZmlyc3RfbmFtZSI6InNhdGlzaCIsIm1pZGRsZV9uYW1lIjoiIiwibGFzdF9uYW1lIjoiam9zaGkiLCJlbWFpbCI6ImR1bW15ZG9jQGdtYWlsLmNvbSIsImdlbmRlciI6Im1hbGUiLCJiaXJ0aF9kYXRlIjoiMTk4NS0wOC0wMSIsInByb2ZpbGVfaW1hZ2UiOiJodHRwczovL2NyeXB0b3BpbGxhcHAuczMuYW1hem9uYXdzLmNvbS9kb2N0b3IvMS9kZW1vLmpwZyIsIm5ld19wcm9maWxlX2ltYWdlIjpudWxsLCJjb250YWN0X251bWJlciI6Iis5MTgxNzg0MjYxNzAiLCJhY2NvdW50X2FjdGl2YXRpb24iOjEsInByb2ZpbGVfaW1hZ2VfdmVyaWZ5IjowLCJsYWJfb3JfcGhhcm1hX2VtcGxveWVtZW50X251bWJlciI6bnVsbCwiZW1haWxfdmVyaWZ5IjowLCJwaG9uZV92ZXJpZnkiOjEsImRlZmF1bHRfcm9sZSI6MiwiaXNBZG1pbiI6MCwiaXNTZXR1cENvbXBsZXRlIjoxLCJkZXZpY2VUb2tlbiI6bnVsbCwiZGV2aWNlVHlwZSI6bnVsbCwicm9sZV9pZCI6Miwicm9sZSI6ImRvY3RvciIsImlzV29ya3BsYWNlQWRtaW4iOjEsImlhdCI6MTYwNTY4MDA5MSwiZXhwIjoxNjA2OTc2MDkxfQ.x35_SNQCcEIRHt2_fmj-8-hxZ3YGeyiLRTt9B87sGrk",
						"default_role": 2,
						"id": 1,
						"isSetupComplete": 1
					},
					"message": "Success"
				}
			},
			"Data2": {
				"title": "Data2",
				"required": ["token", "default_role", "id", "isSetupComplete"],
				"type": "object",
				"properties": {
					"token": {
						"type": "string"
					},
					"default_role": {
						"type": "integer",
						"format": "int32"
					},
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"isSetupComplete": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZmlyc3RfbmFtZSI6InNhdGlzaCIsIm1pZGRsZV9uYW1lIjoiIiwibGFzdF9uYW1lIjoiam9zaGkiLCJlbWFpbCI6ImR1bW15ZG9jQGdtYWlsLmNvbSIsImdlbmRlciI6Im1hbGUiLCJiaXJ0aF9kYXRlIjoiMTk4NS0wOC0wMSIsInByb2ZpbGVfaW1hZ2UiOiJodHRwczovL2NyeXB0b3BpbGxhcHAuczMuYW1hem9uYXdzLmNvbS9kb2N0b3IvMS9kZW1vLmpwZyIsIm5ld19wcm9maWxlX2ltYWdlIjpudWxsLCJjb250YWN0X251bWJlciI6Iis5MTgxNzg0MjYxNzAiLCJhY2NvdW50X2FjdGl2YXRpb24iOjEsInByb2ZpbGVfaW1hZ2VfdmVyaWZ5IjowLCJsYWJfb3JfcGhhcm1hX2VtcGxveWVtZW50X251bWJlciI6bnVsbCwiZW1haWxfdmVyaWZ5IjowLCJwaG9uZV92ZXJpZnkiOjEsImRlZmF1bHRfcm9sZSI6MiwiaXNBZG1pbiI6MCwiaXNTZXR1cENvbXBsZXRlIjoxLCJkZXZpY2VUb2tlbiI6bnVsbCwiZGV2aWNlVHlwZSI6bnVsbCwicm9sZV9pZCI6Miwicm9sZSI6ImRvY3RvciIsImlzV29ya3BsYWNlQWRtaW4iOjEsImlhdCI6MTYwNTY4MDA5MSwiZXhwIjoxNjA2OTc2MDkxfQ.x35_SNQCcEIRHt2_fmj-8-hxZ3YGeyiLRTt9B87sGrk",
					"default_role": 2,
					"id": 1,
					"isSetupComplete": 1
				}
			},
			"UpdatePasswordRequest": {
				"title": "UpdatePassword Request",
				"required": ["id", "password"],
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"password": {
						"type": "string"
					}
				},
				"example": {
					"id": 1,
					"password": "12345678"
				}
			},
			"VerifyOTPRequest": {
				"title": "VerifyOTP Request",
				"required": ["ph_number", "otp"],
				"type": "object",
				"properties": {
					"ph_number": {
						"type": "string"
					},
					"otp": {
						"type": "string"
					}
				},
				"example": {
					"ph_number": "+918178426171",
					"otp": "570605"
				}
			},
			"RequestOTPRequest": {
				"title": "RequestOTP Request",
				"required": [
					"first_name",
					"role_id",
					"middle_name",
					"last_name",
					"contact_number"
				],
				"type": "object",
				"properties": {
					"first_name": {
						"type": "string"
					},
					"role_id": {
						"type": "integer",
						"format": "int32"
					},
					"middle_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"contact_number": {
						"type": "string"
					}
				},
				"example": {
					"first_name": "patient",
					"role_id": 2,
					"middle_name": "",
					"last_name": "user",
					"contact_number": "+918178426171"
				}
			},
			"ChangePasswordRequest": {
				"title": "ChangePasswordRequest",
				"required": [
					"current_password",
					"new_password",
					"confirm_new_password"
				],
				"type": "object",
				"properties": {
					"current_password": {
						"type": "string"
					},
					"new_password": {
						"type": "string"
					},
					"confirm_new_password": {
						"type": "string"
					}
				},
				"example": {
					"current_password": "krishank1234",
					"new_password": "krishank@123",
					"confirm_new_password": "krishank@123"
				}
			},
...
			"ResendOTPRequest": {
				"title": "ResendOTPRequest",
				"required": [
				  "contact_number"
				],
				"type": "object",
				"properties": {
				  "contact_number": {
					"type": "string"
				  }
				},
				"example": {
				  "contact_number": "+918128153228"
				}
			}
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"security": [
		{
			"ApiKeyAuth": []
		}
	]
}